+++ lib/CONFIG/config.cpp
@@ -151,7 +151,11 @@ void Config::fromJson(JsonObject source) {
 }
 
 uint16_t Config::getFrequency() {
-    return conf.frequency;
+    // === TEMPORARY HARDCODE FOR RX5808 CH1 PIN ISSUE ===
+    // Hardcoded to R1 (5658 MHz) - Raceband Channel 1
+    // TODO: Remove this once CH1 pin is fixed and revert to: return conf.frequency;
+    return 5658;
+    // === END TEMPORARY HARDCODE ===
 }
 
 uint32_t Config::getMinLapMs() {
@@ -232,7 +236,7 @@ void Config::setDefaults(void) {
     // Reset everything to 0/false and then just set anything that zero is not appropriate
     memset(&conf, 0, sizeof(conf));
     conf.version = CONFIG_VERSION | CONFIG_MAGIC;
-    conf.frequency = 5917;  // RaceBand Channel 8
+    conf.frequency = 5658;  // RaceBand Channel 1 (R1) - 5658 MHz
     conf.minLap = 50;  // 5.0 seconds
     conf.alarm = 36;
     conf.announcerType = 2;